requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==5.0.1
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-environ==0.11.2
djangorestframework==3.14.0
google-api-core==2.15.0
google-auth==2.26.2
googleapis-common-protos==1.62.0
idna==3.6
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
